Moved tests to test-framework module
